Move CallConnection, IncomingCallConnection, OutgoingCallConnection to separate files
