+++ src/android/MyConnectionService.java
@@ -7,9 +7,7 @@ import org.apache.cordova.PluginResult;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import android.app.Notification;
 import android.app.NotificationChannel;
-import android.app.NotificationManager;
 import android.content.Intent;
 import android.content.Context;
 import android.content.IntentFilter;
@@ -34,21 +32,20 @@ import androidx.annotation.NonNull;
 import androidx.core.app.NotificationCompat;
 import androidx.lifecycle.DefaultLifecycleObserver;
 import androidx.lifecycle.LifecycleOwner;
-import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import java.util.concurrent.ConcurrentHashMap;
 
 public class MyConnectionService extends ConnectionService {
 
     static final String TAG = "MyConnectionService";
-    private static final ConcurrentHashMap<String, Connection> connectionMap = new ConcurrentHashMap<String, Connection>(); // Keys are call_uuid strings
+    static final ConcurrentHashMap<String, Connection> connectionMap = new ConcurrentHashMap<String, Connection>(); // Keys are call_uuid strings
     private static final ConcurrentHashMap<String, Boolean> connectionAddedMap = new ConcurrentHashMap<String, Boolean>(); // Keys are call_uuid strings, true if addIncomingCall called for the given call uuid.
 
     private CallActionReceiver callActionReceiver;
 
     // TODO: store the outgoing connections in connectionMap() to do that we will need a call UUID which we could pass into the app through sendCall
     // we can then get rid of this variable, and just always use connectionMap() + connection UUIDs to access both incoming and outgoing connections.
-    private static Connection activeOutgoingConnection;
+    static Connection activeOutgoingConnection;
 
     @Override
     public void onCreate() {
@@ -141,7 +138,7 @@ public class MyConnectionService extends ConnectionService {
         return START_STICKY; // System will attempt to re-create the service if it is killed.
     }
 
-    private static String activeConnectionUUID;
+    static String activeConnectionUUID;
 
     public static Connection getConnectionByPayload(String pushMessagePayload) {
         JSONObject payload;
@@ -199,7 +196,7 @@ public class MyConnectionService extends ConnectionService {
         }
     }
 
-    private void handleCallAudioStateChanged(CallAudioState state) {
+    void handleCallAudioStateChanged(CallAudioState state) {
         Log.d(TAG, "onCallAudioStateChanged: route=" + state.getRoute() + ", supportedRoutes=" + state.getSupportedRouteMask());
 
         // Use the centralized method from CordovaCall to emit route change event
@@ -233,7 +230,7 @@ public class MyConnectionService extends ConnectionService {
 
         connectionAddedMap.remove(callUUID);
 
-        final IncomingCallConnection connection = new IncomingCallConnection(callUUID, payloadString, callerName);
+        final IncomingCallConnection connection = new IncomingCallConnection(this, callUUID, payloadString, callerName);
 
         connection.setCallerDisplayName(callerName, TelecomManager.PRESENTATION_ALLOWED);
 
@@ -262,7 +259,7 @@ public class MyConnectionService extends ConnectionService {
 
     @Override
     public Connection onCreateOutgoingConnection(PhoneAccountHandle connectionManagerPhoneAccount, ConnectionRequest request) {
-        final OutgoingCallConnection connection = new OutgoingCallConnection();
+        final OutgoingCallConnection connection = new OutgoingCallConnection(this);
         connection.setAddress(Uri.parse(request.getExtras().getString("to")), TelecomManager.PRESENTATION_ALLOWED);
         Icon icon = CordovaCall.getIcon();
         if(icon != null) {
@@ -290,157 +287,5 @@ public class MyConnectionService extends ConnectionService {
 
         activeOutgoingConnection = connection;
         return connection;
-    }
-
-    /**
-     * Base class for both incoming and outgoing connections.
-     * Holds the common logic shared between IncomingCallConnection and OutgoingCallConnection.
-     */
-    private class CallConnection extends Connection {
-        @Override
-        public void onCallAudioStateChanged(CallAudioState state) {
-            super.onCallAudioStateChanged(state);
-            handleCallAudioStateChanged(state);
-        }
-
-        @Override
-        public void onAbort() {
-            this.setDisconnected(new DisconnectCause(DisconnectCause.CANCELED));
-        }
-
-        @Override
-        public void onDisconnect() {
-            this.setDisconnected(new DisconnectCause(DisconnectCause.LOCAL));
-            CordovaCall.emitEvent("hangup", new PluginResult(PluginResult.Status.OK, "hangup event called successfully"));
-        }
-
-        @Override
-        public void onStateChanged(int state) {
-            super.onStateChanged(state);
-            if (state == Connection.STATE_ACTIVE) {
-                CordovaCall.onCallConnected();
-            } else if (state == Connection.STATE_DISCONNECTED) {
-                this.destroy();
-                CordovaCall.onCallEnded();
-                Log.d(TAG, "Stopping CallAudioService...");
-                Context context = getApplicationContext();
-                Intent serviceIntent = new Intent(context, CallAudioService.class);
-                context.stopService(serviceIntent);
-            }
-        }
-    }
-
-    private class IncomingCallConnection extends CallConnection {
-        private final String callUUID;
-        private final String payloadString;
-        private final String callerName;
-        private IncomingCallNotification incomingCallNotification;
-        private Runnable mainActivityChangeListener;
-
-        IncomingCallConnection(String callUUID, String payloadString, String callerName) {
-            this.callUUID = callUUID;
-            this.payloadString = payloadString;
-            this.callerName = callerName;
-        }
-
-        @Override
-        public void onShowIncomingCallUi() { // Only for self managed connections
-            Log.d(TAG, "onShowIncomingCallUi() invoked, for call_uuid: " + callUUID);
-            this.setRinging();
-
-            this.incomingCallNotification = new IncomingCallNotification(payloadString, getApplicationContext());
-            Notification notification = this.incomingCallNotification.build();
-
-            NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-            notificationManager.notify(this.incomingCallNotification.getNotificationID(), notification);
-        }
-
-        public void cancelIncomingCallNotification() {
-            if (this.incomingCallNotification == null) {
-                return;
-            }
-            NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-            notificationManager.cancel(this.incomingCallNotification.getNotificationID());
-            this.incomingCallNotification = null;
-        }
-
-        @Override
-        public void onAnswer() {
-            Log.d(TAG, "onAnswer()");
-
-            cancelIncomingCallNotification();
-
-            this.setActive();
-            activeConnectionUUID = callUUID;
-
-            showWebApp("answerCall", payloadString);
-
-            Log.d(TAG, "Starting CallAudioService...");
-            Intent intent = new Intent(getApplicationContext(), CallAudioService.class);
-            intent.putExtra("peerName", callerName);
-            startForegroundService(intent);
-
-            Log.d(TAG, "Emitting CordovaCall answer event...");
-            CordovaCall.emitEvent("answer", new PluginResult(PluginResult.Status.OK, payloadString));
-        }
-
-        @Override
-        public void onReject() {
-            Log.d(TAG, "onReject, call_uuid: " + callUUID);
-            this.setDisconnected(new DisconnectCause(DisconnectCause.REJECTED));
-
-            showWebApp("declineCall", payloadString); // Controversial UX but doing so that we can tell the web app to reject the call (which may let the caller know it was declined)
-
-            CordovaCall.emitEvent("reject", new PluginResult(PluginResult.Status.OK, payloadString));
-        }
-
-        @Override
-        public void onAbort() {
-            Log.d(TAG, "onAbort, call_uuid: " + callUUID);
-            super.onAbort();
-        }
-
-        @Override
-        public void onDisconnect() {
-            Log.d(TAG, "onDisconnect, call_uuid: " + callUUID);
-            super.onDisconnect();
-        }
-
-        @Override
-        public void onStateChanged(int state) {
-            Log.d(TAG, "connection onStateChanged: " + state);
-
-            if (state == Connection.STATE_DISCONNECTED) {
-                connectionMap.remove(callUUID);
-                if (activeConnectionUUID != null && activeConnectionUUID.equals(callUUID)) {
-                    activeConnectionUUID = null;
-                }
-                if (this.mainActivityChangeListener != null) {
-                    CordovaCall.unregisterMainActivityStateChangeListener(this.mainActivityChangeListener);
-                }
-                cancelIncomingCallNotification();
-            }
-
-            super.onStateChanged(state);
-
-            Log.d(TAG, "broadcasting connection_state_changed call_uuid: " + callUUID + " state: " + state);
-            Intent intent = new Intent("connection_state_changed");
-            intent.putExtra("call_uuid", callUUID);
-            intent.putExtra("state", state);
-            LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
-        }
-    }
-
-    private class OutgoingCallConnection extends CallConnection {
-        @Override
-        public void onStateChanged(int state) {
-            if (state == Connection.STATE_DISCONNECTED) {
-                // In all cases when connection transitions to STATE_DISCONNECTED (both onAbort() and onDisconnect())
-                // Ensure the connection is destroyed, etc.
-                activeOutgoingConnection = null;
-                activeConnectionUUID = null;
-            }
-            super.onStateChanged(state);
-        }
     }
 }
